ISSUE-1788 Extract Feign RabbitMQManagementAPI to a common module
To be reused in TWP settings consumer healthcheck
